requirements.txt
# AI and APIs
openai>=1.0.0

# Audio processing
pydub>=0.25.1
soundfile>=0.12.1

# Video processing
moviepy>=1.0.3
opencv-python>=4.8.0

# Image processing
Pillow>=10.0.0

# Web scraping
duckduckgo-search>=3.9.0
requests>=2.31.0

# Email
yagmail>=0.15.0

# Configuration
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
